refactor(mobile): extract coding agent thread actions
