+++ specs/105-coding-agent-shells/current-state.md
@@ -321,6 +321,7 @@ Screen:
 - Approval-request events render allowed decisions in the mobile thread route. Decisions use mobile-generated idempotency request ids, go through the authenticated gateway client, and replace local thread details with the gateway-returned bounded snapshot. Failed decisions show a generic recovery-oriented message.
 - User-input request events render a transient bounded answer composer in the mobile thread route. Answers use mobile-generated idempotency request ids, go through the authenticated gateway client, and replace local thread details with the gateway-returned bounded snapshot. Failed submissions show a generic recovery-oriented message.
 - Successful mobile approval decisions and user-input answers trigger local success haptics only after the gateway-returned bounded snapshot is accepted; haptic failures are logged and do not block the thread state update.
+- Mobile approval/input action state is isolated in `apps/mobile/lib/agent-thread-actions.ts`; the route owns hydration, streaming, and composition while the hook owns transient pending ids, safe per-action errors, bounded input drafts, idempotency request ids, and accepted-snapshot haptic guards.
 - Thread details and Recent Work terminal rows can open the existing `/terminal` tab after persisting only the safe canonical shell-session reference in `mobile-shell-state`; terminal output and transcripts remain outside AsyncStorage.
 - Thread details can open the existing mobile composer for a follow-up run with only bounded source thread id/title/provider route params. The composer validates those params with shared contract schemas and sends a `structured_ref` attachment for the source thread through the normal authenticated create-thread gateway client.
 - Review-ready events in the mobile thread timeline can open the Agents review section with only a bounded `reviewId` route param. The Agents screen validates that route param with `ReviewIdSchema`, rehydrates the review snapshot through the authenticated gateway client, and ignores invalid values without persisting review or diff content.
